Move adapter module into fitters module
