+++ src/prfmodel/fitters/adapter.py
